requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.31.0
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
